Changes goaling to make it available to compile the library with Emscripten. One big change is added code for OpenGL ES 3.0 and switching between OpenGL 4.5 and OpenGL ES 3.0 basing on current compilation platform.

+++ libraries/Tilc/Game2D/World/MapLayer.cpp
@@ -107,7 +107,7 @@ void Tilc::Game2D::TMapLayer::LoadFromStringSimplified(TTileSet& TileSet, const 
 		}
 		else if (m_MapWidth != W)
 		{
-			SDL_Log("Distinct length row found on position %llu", x);
+			SDL_Log("Distinct length row found on position %llu", static_cast<unsigned long long>(x));
 			m_MapWidth = 0;
 			break;
 		}
